requirements.txt
numpy==1.24.4
pyzmq==25.1.1
torch==2.1.0
torchvision==0.16.0
